+++ click_extra/tests/conftest.py
@@ -199,3 +199,23 @@ default_options_colored_help = (
     r"  \x1b\[36m--version\x1b\[0m                 Show the version and exit.\x1b\[0m\n"
     r"  \x1b\[36m-h, \x1b\[36m--help\x1b\[0m\x1b\[0m                Show this message and exit.\x1b\[0m\n"
 )
+
+
+default_debug_uncolored_log = (
+    r"debug: Verbosity set to DEBUG.\n"
+    r"debug: Search for configuration in default location...\n"
+    r"debug: No default configuration found.\n"
+    r"debug: No configuration provided.\n"
+    r"debug: \S+, version \S+\n"
+    r"debug: {.*}\n"
+)
+
+
+default_debug_colored_log = (
+    r"\x1b\[34mdebug: \x1b\[0mVerbosity set to DEBUG.\n"
+    r"\x1b\[34mdebug: \x1b\[0mSearch for configuration in default location...\n"
+    r"\x1b\[34mdebug: \x1b\[0mNo default configuration found.\n"
+    r"\x1b\[34mdebug: \x1b\[0mNo configuration provided.\n"
+    r"\x1b\[34mdebug: \x1b\[0m\x1b\[97m\S+\x1b\[0m, version \x1b\[32m\S+\x1b\[0m\x1b\[90m\n"
+    r"\x1b\[34mdebug: \x1b\[0m{.*}\x1b\[0m\n"
+)